Initial Maven module structure, closes #8
