+++ modules/core/src/main/kotlin/org/coner/trailer/Participant.kt
