+++ modules/core/src/main/kotlin/org/coner/trailer/Person.kt
