+++ modules/core/src/main/kotlin/org/coner/trailer/eventresults/ResultRun.kt
