+++ modules/core/src/main/kotlin/org/coner/trailer/eventresults/ResultsReport.kt
