+++ modules/core/src/main/kotlin/org/coner/trailer/eventresults/ResultsType.kt
