+++ modules/core/src/main/kotlin/org/coner/trailer/seasonpoints/ParticipantResultPointsCalculator.kt
