+++ modules/core/src/main/kotlin/org/coner/trailer/seasonpoints/PersonStandingAccumulator.kt
