+++ modules/core/src/test/kotlin/org/coner/trailer/TestEvents.kt
