+++ modules/core/src/test/kotlin/org/coner/trailer/TestPeople.kt
