+++ modules/core/src/test/kotlin/org/coner/trailer/eventresults/MockkParticipantResult.kt
