+++ modules/core/src/test/kotlin/org/coner/trailer/eventresults/MockkResultRun.kt
