+++ modules/core/src/test/kotlin/org/coner/trailer/eventresults/ParticipantResultTest.kt
