+++ modules/core/src/test/kotlin/org/coner/trailer/seasonpoints/StandingsReportCreatorTest.kt
