--- src/main/kotlin/org/coner/trailer/eventresults/BestRunParticipantResult.kt
@@ -1,4 +1,0 @@
-package org.coner.trailer.eventresults
-
-import org.coner.trailer.Participant
-
